Fetch images for thumbnails asynchronously
