+++ app/storage.py
@@ -4,6 +4,7 @@ import tempfile
 import uuid
 import pathlib
 
+import gevent
 import libcloud.storage.types
 import magic
 from mutagen.mp4 import MP4
@@ -269,6 +270,7 @@ def calculate_file_hash(ufile, size_limit=None):
         if size_limit is not None and size > size_limit:
             raise SizeLimitExceededError
         fhash.update(data)
+        gevent.sleep(0)
     return fhash.hexdigest()
